Add timer.stop to receive script
Added a tmr.stop(0) command to the receive script. To prevent boot
loops, apps should be launched with tmr.delay() for a few seconds to
allow a time to regain control of the MCU by stopping the timer. This
change takes advantage of that to prevent the MCU from 'booting'
completely so it can receive the file instead.